better workflow tables

+++ webEdition/liveUpdate/sqldumps/tblWorkflowLog.sql
@@ -2,7 +2,7 @@ CREATE TABLE ###TBLPREFIX###tblWorkflowLog (
   ID mediumint unsigned NOT NULL auto_increment,
   RefID int unsigned NOT NULL default '0',
   docTable enum('tblWorkflowLog') NOT NULL default 'tblWorkflowLog',
-  userID int unsigned NOT NULL default '0',
+  userID smallint unsigned NOT NULL default '0',
   logDate int unsigned NOT NULL default '0',
   `Type` tinyint unsigned NOT NULL default '0',
   Description tinytext NOT NULL,
