+++ webEdition/liveUpdate/sqldumps/tblWorkflowTask.sql
@@ -1,8 +1,8 @@
 CREATE TABLE ###TBLPREFIX###tblWorkflowTask (
   ID int unsigned NOT NULL auto_increment,
-  userID int unsigned NOT NULL default '0',
-  Edit int unsigned NOT NULL default '0',
-  Mail int unsigned NOT NULL default '0',
+  userID mediumint unsigned NOT NULL default '0',
+  Edit tinyint unsigned NOT NULL default '0',
+  Mail tinyint unsigned NOT NULL default '0',
   stepID int unsigned NOT NULL default '0',
   PRIMARY KEY  (ID)
 ) ENGINE=MyISAM;
